move resource index functions to search app and begin order by
