+++ physionet-django/physionet/views.py
@@ -65,50 +65,3 @@ def contact(request):
 
 def citi_instructions(request):
     return render(request, 'about/citi_instructions.html')
-
-# Content pages
-
-
-def get_content(resource_type=None):
-    """
-    Helper function to get content shown on a resource listing page
-    """
-    if resource_type is None:
-        published_projects = PublishedProject.objects.all().order_by(
-            '-publish_datetime')
-    else:
-        published_projects = PublishedProject.objects.filter(
-            resource_type=resource_type).order_by('-publish_datetime')
-
-    authors = [p.authors.all() for p in published_projects]
-    topics = [p.topics.all() for p in published_projects]
-    projects_authors_topics = zip(published_projects, authors, topics)
-
-    return projects_authors_topics
-
-
-def content_index(request):
-    """
-    List of all published resources
-    """
-    projects_authors_topics = get_content()
-    return render(request, 'content_index.html', {'content_name':'Content',
-        'projects_authors_topics':projects_authors_topics})
-
-
-def database_index(request):
-    """
-    List of published databases
-    """
-    projects_authors_topics = get_content(resource_type=0)
-    return render(request, 'database_index.html', {'content_name':'Database',
-        'projects_authors_topics':projects_authors_topics})
-
-
-def software_index(request):
-    """
-    List of published software projects
-    """
-    projects_authors_topics = get_content(resource_type=1)
-    return render(request, 'software_index.html', {'content_name':'Software',
-        'projects_authors_topics':projects_authors_topics})
